--- a/src/java.base/share/classes/sun/security/util/math/intpoly/IntegerPolynomial.java
+++ b/src/java.base/share/classes/sun/security/util/math/intpoly/IntegerPolynomial.java
@@ -824,7 +824,7 @@ public abstract sealed class IntegerPolynomial implements IntegerFieldModuloP
             }
 
             IntegerPolynomial.this.square(limbs, limbs);
-            numAdds = 0; 
+            numAdds = 0;
             return this;
         }