@vnkozlov 2024-06-17
What causes regression in P256 "(~-8-14%)"?
From what I see, you re-arranged code to not execute some code ("reducePositive()") when it is not needed. How this affects P256?