@vnkozlov 2024-06-17
> Actually, the other way around; reducePositive is now an unconditionally executed for both pure java and the intrinsic paths. 

Looking on `MontgomeryIntegerPolynomialP256.java` the code in `multImpl() + reducePositive()` is similar to original `mult()` except new additional code at the end of `multImpl()`. Now you intrinsify only `multImpl()`. Looks like `reducePositive()`is not included into intrinsic and will be normally JIT compiled (hopeful inlined when JIT compiling `mult()`?

Also you did not change assembler for intrinsic but you changed corresponding Java code (`multImpl()`).  How it works?