@vnkozlov 2024-06-17
Talking about further improvements. Is it possible to optimize reduction code by converting it to intrinsic too?  Or code generated by C2 is good enough?