@ferakocz 2024-06-25
Looks good to me. It would be good, though, to figure out what else could be done to regain the P256 performance with keeping the speed of this code path.