@vpaprotsk 2024-06-25
/backport :jdk23